+++ src/commands/build-web-app.js
@@ -1,9 +1,13 @@
 import path from 'path'
 
+import {copyPublicDir} from '../utils'
+
 import webpackBuild from '../webpackBuild'
 
 export default function(args, cb) {
   require('./clean-app')(args)
+
+  copyPublicDir('public', 'dist')
 
   console.log(`nwb: build-web-app`)
   webpackBuild(args, {
@@ -13,10 +17,13 @@ export default function(args, cb) {
     },
     output: {
       filename: '[name].js',
-      path: path.resolve('public/build'),
-      publicPath: '/build/'
+      path: path.resolve('dist'),
+      publicPath: '/'
     },
     plugins: {
+      html: {
+        template: path.resolve('src/index.html')
+      },
       vendorChunkName: 'vendor'
     }
   }, cb)