+++ src/commands/clean-app.js
@@ -2,5 +2,5 @@ import exec from '../exec'
 
 export default function(args) {
   console.log('nwb: clean-app')
-  exec('rimraf', ['coverage', 'public/build'])
+  exec('rimraf', ['coverage', 'dist'])
 }