+++ src/devServer.js
@@ -1,7 +1,5 @@
-import fs from 'fs'
-import path from 'path'
-
 import chalk from 'chalk'
+import history from 'connect-history-api-fallback'
 import express from 'express'
 import webpack from 'webpack'
 
@@ -10,12 +8,15 @@ import webpack from 'webpack'
  * assets using Webpack's watch mode, and webpack-hot-middleware to hot reload
  * changes in the browser.
  *
- * If static path config is also provided, express will also be used to serve
- * static content from the given path.
+ * If static path config is provided, express will serve static content from it.
  */
 export default function server(webpackConfig, {fallback, noInfo, port, staticPath}, cb) {
   let app = express()
   let compiler = webpack(webpackConfig)
+
+  if (fallback) {
+    app.use(history())
+  }
 
   app.use(require('webpack-dev-middleware')(compiler, {
     noInfo,
@@ -29,21 +30,6 @@ export default function server(webpackConfig, {fallback, noInfo, port, staticPat
 
   if (staticPath) {
     app.use(express.static(staticPath))
-  }
-
-  if (fallback) {
-    app.use((req, res, next) => {
-      // Only fall back for GET methods which accept HTML and don't appear to
-      // end with a file extension.
-      if (req.method !== 'GET' || !req.accepts('html') || /\.[\w]{1,4}$/i.test(req.path)) {
-        return next()
-      }
-      fs.readFile(path.resolve('public/index.html'), 'utf8', (err, html) => {
-        if (err) return next(err)
-        // Rewrite relative build URLs to be absolute so they work from any path
-        res.send(html.replace(/="build\//g, '="/build/'))
-      })
-    })
   }
 
   app.listen(port, 'localhost', err => {